+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Version [1.0.0](https://github.com/cedx/koa-eta/compare/v0.2.0...v1.0.0)
 - First stable release.
+- Added an example of a [Koa](https://koajs.com) server.
 
 ## Version [0.2.0](https://github.com/cedx/koa-eta/compare/v0.1.2...v0.2.0)
 - Add the `async` rendering option.